requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
python-chess>=1.9.0

# Optional but recommended
h5py>=3.8.0        # HDF5 caching for pretraining
cairosvg>=2.7.0    # SVG rendering for chess pieces
pillow>=10.0.0     # Image processing for UI

# Note: tkinter is usually included with Python
# If not, install via your system package manager:
#   Ubuntu/Debian: sudo apt-get install python3-tk
#   Fedora: sudo dnf install python3-tkinter
#   macOS: brew install python-tk
